move debug utils to common
