+++ player/css/player.css
@@ -45,6 +45,7 @@
     display: block !important;
     visibility: visible !important;
     opacity: 1 !important;
+    border: 1px solid white !important; /* Add white border */
 }
 
 /* Progress bar */
