+++ player/js/ui/ProgressBar.js
@@ -1,6 +1,6 @@
 // Disclaimer: no semicolons, if unnecessary, are used in this project
 
-import { debugLog } from '../utils/debug-utils.js'
+import { debugLog } from '../../../common/utils/debug-utils.js' // Updated path
 
 export class ProgressBar {
     constructor(parent, options = {}) {
