+++ player/js/utils/player-utils.js
@@ -1,5 +1,7 @@
 // Disclaimer: no semicolons, if unnecessary, are used in this project
 
+import { debugLog } from '../../../common/utils/debug-utils.js' // Updated path
+
 export function getMediaListContainer(element) {
     let current = element
     while (current && !current.mediaList) {
